requirements.txt
attrs==21.4.0
coverage==6.3.2
cycler==0.11.0
docxcompose==1.3.4
docxtpl==0.15.2
fonttools==4.29.1
iniconfig==1.1.1
Jinja2==3.0.3
kiwisolver==1.3.2
lxml==4.8.0
MarkupSafe==2.1.0
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pyqt-darktheme==1.2.5
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyQt5-stubs==5.15.2.0
pytest==7.0.1
python-dateutil==2.8.2
python-docx==0.8.11
six==1.16.0
tomli==2.0.1
